requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
pandas>=1.5.0
pyyaml>=6.0
flask>=2.3.0
lxml>=4.9.0 